+++ etc/install.bash
