--- launch/polhemus.launch
@@ -1,5 +1,0 @@
-<launch>
-
-	<include file="$(find polhemus_ros)/launch/polhemus.launch.xml" />
-
-</launch>
