+++ package.xml
@@ -1,12 +1,13 @@
 <?xml version="1.0"?>
-<package>
-  <name>polhemus_ros</name>
-  <version>0.0.0</version>
-  <description>The polhemus_ros package</description>
+<package format="2">
+  <name>polhemus_ros_driver</name>
+  <version>0.0.1</version>
+  <description>The polhemus ROS driver package</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag --> 
   <!-- Example:  -->
   <!-- <maintainer email="jane.doe@example.com">Jane Doe</maintainer> -->
+  <maintainer email="chris.mower@ed.ac.uk">chris mower</maintainer>
   <maintainer email="yiming@todo.todo">yiming</maintainer>
 
 
@@ -28,35 +29,10 @@
   <!-- <author email="jane.doe@example.com">Jane Doe</author> -->
 
 
-  <!-- The *_depend tags are used to specify dependencies -->
-  <!-- Dependencies can be catkin packages or system dependencies -->
-  <!-- Examples: -->
-  <!-- Use build_depend for packages you need at compile time: -->
-  <!--   <build_depend>message_generation</build_depend> -->
-  <!-- Use buildtool_depend for build tool packages: -->
-  <!--   <buildtool_depend>catkin</buildtool_depend> -->
-  <!-- Use run_depend for packages you need at runtime: -->
-  <!--   <run_depend>message_runtime</run_depend> -->
-  <!-- Use test_depend for packages you need only for testing: -->
-  <!--   <test_depend>gtest</test_depend> -->
   <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>geometry_msgs</build_depend>
-  <build_depend>tf</build_depend>
-  <build_depend>roscpp</build_depend>
-  <build_depend>kdl_parser</build_depend>
-  <build_depend>exotica</build_depend>
-  <run_depend>geometry_msgs</run_depend>
-  <run_depend>tf</run_depend>
-  <run_depend>roscpp</run_depend>
-  <run_depend>kdl_parser</run_depend>
-  <run_depend>exotica</run_depend>
 
-  <!-- The export tag contains other, unspecified, tags -->
-  <export>
-    <!-- You can specify that this package is a metapackage here: -->
-    <!-- <metapackage/> -->
-
-    <!-- Other tools can request additional information be placed here -->
-
-  </export>
+  <depend>tf2</depend>
+  <depend>tf2_ros</depend>
+  <depend>roscpp</depend>
+  <depend>geometry_msgs</depend>
 </package>
