--- readme.md
@@ -1,6 +1,0 @@
-# Polhemus Magnetic Motion tracker driver
-
-This package depends on libusb-dev and fxload libraries. To install these and to set up the udev rule run:
-1. ```cd etc```
-2. ```./setup.bash```
-The setup script will ask for sudo password.
